Initial controller firmware.
This implements receiving the infrared RC5 signal and decodes it, then toggles
the appropriate GPIO pins to output the correct levels to the player. A delay
mechanism has been implemented to give the output the necessary minimum time to
get recognized by the player.

This code assumes an updated circuit to fix the power on issue. Verified with a
player running Rockbox, the Iriver firmware might need some timing adjustments.